requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.27.0.post1
gunicorn==23.0.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart
upstash-redis>=1.0
# YOLO 및 이미지 처리
ultralytics==8.3.50
opencv-python==4.10.0.84
numpy==1.26.4
redis>=5.0